requirements.txt
newspaper3k
pyqt5
requests
yake
lxml_html_clean
selenium